+++ stock_comber/datasources/__init__.py
@@ -1,8 +1,11 @@
 """Free, key-less online data sources for fundamentals and prices."""
 
 from .cache import FileCache
+from .finnhub import FinnhubSource
 from .sec_edgar import SecEdgarSource
 from .stooq import StooqSource
 from .yahoo import YahooSource
 
-__all__ = ["FileCache", "SecEdgarSource", "StooqSource", "YahooSource"]
+__all__ = [
+    "FileCache", "FinnhubSource", "SecEdgarSource", "StooqSource", "YahooSource",
+]